Allow running `wrangler dev` without an account_id
Previously, running it with no account_id and no config in ~/.wrangler would give an error:

```
Error: config path does not exist /home/jnelson/.wrangler/config/default.toml. Try running `wrangler login` or `wrangler config`
```

This is not technically a regression from loading account_id lazily, because in wrangler 1.17 it would just error earlier that account_id was required:
```
🕵️  You can find your account_id in the right sidebar of your account's Workers page
Error: field `account_id` is required to deploy to workers.dev
```

However, it seems odd that you'd have to have an account to run an unauthenticated preview.
This changes wrangler to only require an account_id or config file if you're actually deploying your site.

This also changes the tests not to expect that wrangler requires an account_id just to load the available deployments.